backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic[email]
email-validator
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
authlib
httpx
itsdangerous
sendgrid
alembic

mistralai==1.0.0
gitpython
pytest-json-report
flake8